+++ Changelog.md
@@ -6,6 +6,10 @@ Enhancements:
 * Improve pluralisation of words ending with `s` (like process). (Joshua Pinter, #2779)
 * Add ordering by file modification time (most recent first). (Matheus Richard, #2778)
 
+Bug fixes:
+
+* Ensure bisect communication uses consistent encoding. (Mike Jarema, #2852)
+
 ### 3.10.1 / 2020-12-27
 [Full Changelog](http://github.com/rspec/rspec-core/compare/v3.10.0...v3.10.1)