+++ Changelog.md
@@ -1,3 +1,10 @@
+### Development
+[Full Changelog](http://github.com/rspec/rspec-core/compare/v3.10.1...3-10-maintenance)
+
+Bug fixes:
+
+* Ensure bisect communication uses consistent encoding. (Mike Jarema, #2852)
+
 ### 3.10.1 / 2020-12-27
 [Full Changelog](http://github.com/rspec/rspec-core/compare/v3.10.0...v3.10.1)